// src/interfaces/IPokedex.ts
import { IApiResource } from './IApiResource';
// Generated by https://quicktype.io

export interface IPokedex {
  descriptions: Description[];
  id: number;
  is_main_series: boolean;
  name: string;
  names: Name[];
  pokemon_entries: PokemonEntry[];
  region: null;
  version_groups: string[];
}

export interface Description {
  description: string;
  language: IApiResource;
}

export interface Name {
  language: IApiResource;
  name: string;
}

export interface PokemonEntry {
  entry_number: number;
  pokemon_species: IApiResource;
}